Simulated detector and sample selector work.
