+++ utils/simulated_hardware.py
@@ -1,0 +1,49 @@
+import threading
+
+import numpy
+import numpy as np
+from ophyd import Component, Device, Signal, DeviceStatus
+from bluesky.plan_stubs import mv
+
+from .generate_data import generate_measured_image, SHAPE
+
+
+sample_selector = Signal(value=0, name="sample_selector")
+
+class TimerStatus(DeviceStatus):
+    """Simulate the time it takes for a detector to acquire an image."""
+
+    def __init__(self, device, delay):
+        super().__init__(device)
+        self.delay = delay  # for introspection purposes
+        threading.Timer(delay, self.set_finished).start()
+
+
+class DiffractionDetector(Device):
+    # exposure_time = Component(Signal, value=1)
+    image = Component(Signal, value=numpy.zeros(SHAPE))
+    signal_to_noise = Component(Signal, value=0)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def trigger(self):
+        "Generate a simulated reading with noise for the current sample."
+        sample_number = sample_selector.value
+        arr = generate_measured_image(sample_number)
+        # Update the internal signal with a simulated image.
+        self.image.set(arr)
+        # Simulate the exposure and readout time.
+        EXPOSURE_TIME = 2
+        READOUT_TIME = 0.17
+        return TimerStatus(self, EXPOSURE_TIME + READOUT_TIME)
+
+    def collect_asset_docs(self):
+        yield from []
+
+
+detector = DiffractionDetector(name="detector")
+
+
+def select_sample(sample_number):
+    yield from mv(sample_selector, sample_number)
